+++ test/TestContracts/PositionManagerTester.sol
@@ -2,7 +2,10 @@
 
 pragma solidity 0.8.19;
 
-import "../../contracts/PositionManager.sol";
+import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import {IPriceFeed} from "../../contracts/Interfaces/IPriceFeed.sol";
+import {PositionManager} from "../../contracts/PositionManager.sol";
+import {MathUtils} from "../../contracts/Dependencies/MathUtils.sol";
 
 /* Tester contract inherits from PositionManager, and provides external functions
 for testing the parent's internal functions. */
@@ -14,24 +17,25 @@ contract PositionManagerTester is PositionManager {
         uint256 _positionsSize,
         uint256 _liquidationProtocolFee,
         address[] memory delegates
-    )
-        PositionManager(_priceFeed, _collateralToken, _positionsSize, _liquidationProtocolFee, delegates)
-    {
-    }
+    ) PositionManager(_priceFeed, _collateralToken, _positionsSize, _liquidationProtocolFee, delegates) {}
 
-    function getCollGasCompensation(uint _coll) external pure returns (uint) {
+    function getCollGasCompensation(uint256 _coll) external pure returns (uint256) {
         return MathUtils.getCollGasCompensation(_coll);
     }
 
-    function getCollLiquidationProtocolFee(uint _entireColl, uint _entireDebt, uint _price, uint _fee) external pure returns (uint) {
+    function getCollLiquidationProtocolFee(uint256 _entireColl, uint256 _entireDebt, uint256 _price, uint256 _fee)
+        external
+        pure
+        returns (uint256)
+    {
         return _getCollLiquidationProtocolFee(_entireColl, _entireDebt, _price, _fee);
     }
 
-    function getRGasCompensation() external pure returns (uint) {
+    function getRGasCompensation() external pure returns (uint256) {
         return MathUtils.R_GAS_COMPENSATION;
     }
 
-    function unprotectedDecayBaseRateFromBorrowing() external returns (uint) {
+    function unprotectedDecayBaseRateFromBorrowing() external returns (uint256) {
         baseRate = _calcDecayedBaseRate();
         assert(baseRate >= 0 && baseRate <= MathUtils._100_PERCENT);
 
@@ -43,11 +47,11 @@ contract PositionManagerTester is PositionManager {
         lastFeeOperationTime = block.timestamp;
     }
 
-    function setBaseRate(uint _baseRate) external {
+    function setBaseRate(uint256 _baseRate) external {
         baseRate = _baseRate;
     }
 
-    function getActualDebtFromComposite(uint _debtVal) external pure returns (uint) {
+    function getActualDebtFromComposite(uint256 _debtVal) external pure returns (uint256) {
         return MathUtils.getNetDebt(_debtVal);
     }
 }
